+++ jsapp/js/components/library/collectionAssetsTable.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import AssetsTable from 'js/components/assetsTable/assetsTable';
 import {AssetsTableContextName} from 'js/components/assetsTable/assetsTableConstants';
-import type {OrderDirection} from 'js/projects/projectsView/projectsViewConstants';
+import type {OrderDirection} from 'js/projects/projectViews/constants';
 import singleCollectionStore from './singleCollectionStore';
 import type {AssetResponse} from 'js/dataInterface';
