+++ jsapp/js/components/library/myLibraryStore.ts
@@ -12,7 +12,7 @@ import {
   ORDER_DIRECTIONS,
   ASSETS_TABLE_COLUMNS,
 } from 'js/components/assetsTable/assetsTableConstants';
-import type {OrderDirection} from 'js/projects/projectsView/projectsViewConstants';
+import type {OrderDirection} from 'js/projects/projectViews/constants';
 import {ROUTES} from 'js/router/routerConstants';
 import { history } from "js/router/historyRouter";
 import type {
