+++ jsapp/js/projects/projectViews/viewSwitcher.tsx
