+++ jsapp/js/projects/projectsTable/projectsTable.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import InfiniteScroll from 'react-infinite-scroller';
 import LoadingSpinner from 'js/components/common/loadingSpinner';
 import ProjectsTableRow from './projectsTableRow';
-import type {ProjectFieldName, OrderDirection} from 'js/projects/projectsView/projectsViewConstants';
-import {PROJECT_FIELDS} from 'js/projects/projectsView/projectsViewConstants';
+import type {ProjectFieldName, OrderDirection} from 'js/projects/projectViews/constants';
+import {PROJECT_FIELDS} from 'js/projects/projectViews/constants';
 import ProjectsTableHeader from './projectsTableHeader';
 import type {AssetResponse} from 'js/dataInterface';
 import styles from './projectsTable.module.scss';
